// backend/src/middleware/auth.d.ts
import { Request, Response, NextFunction } from 'express';
interface AuthRequest extends Request {
    userId?: string;
}
export declare const authenticateToken: (req: AuthRequest, res: Response, next: NextFunction) => Response<any, Record<string, any>> | undefined;
export {};
//# sourceMappingURL=auth.d.ts.map